fixed multiple bugs introduced a while ago
- just changing everything to uin64_t was a spectacularily stupid idea
- use appropiate GL types for all sizes, offsets, etc.
- make mesh_dump output buffer contents
- fix bug mesh->index buffer would not by resized correctly

+++ io.c
@@ -16,7 +16,7 @@
 
 #include "io.h"
 
-char* read_file(const char* filename, uint64_t* length) {
+char* read_file(const char* filename, int32_t* length) {
     FILE* file = fopen(filename, "rb");
 
     if( ! file ) {
